Adjusted what exceptions can be thrown in the getBalance method.
Adjusted the getBalance methods in the AddressAPI interface, the
AbsBockchains abstract class and the Blockchain class so that it doesn't
throw MalformedURLException and InvalidURIException. Instead, these
exceptions are caught within the Blockchain class and in turn throws an
IOException as this is where the URL/URI is created, and as such, this
is where incorrectly created URI's should be handled.